Fix audioplayer in windows with milisecond level updates

+++ packages/audioplayers_windows/windows/audio_player.h
@@ -69,7 +69,7 @@ public:
     bool GetLooping();
     int64_t GetPosition();
     int64_t GetDuration();
-    void SeekTo(int64_t seek);
+    void SeekTo(double seek);
 
     void SetSourceUrl(std::string url);
